requirements.txt
streamlit==1.54.0
google-generativeai==0.8.6
python-dotenv==1.2.1
langchain-huggingface==1.2.0
PyPDF2==3.0.1
langchain==1.2.10
langchain-text-splitters==1.1.0
google-genai==1.62.0                                                                                                                                              
langchain-google-genai==4.2.0
langchain_community==0.4.1
langchain-classic==1.0.1
pdf2image==1.17.0
sentence-transformers==5.2.2
faiss-cpu==1.13.2